Pass BAO_ADDR to the token helper
Previously the token helper might inherit BAO_ADDR from the process,
but if the address was specified through an -address command-line flag,
then the token helper would not know the address, or it would use the
wrong one. Fix that by propagating the address everywhere, and then
setting BAO_ADDR explicitly in the token helper's environment.

Fixes #314.

Signed-off-by: Ruud van Asseldonk <ruud@chorus.one>