Add support to Transit, keysutil for XChaCha20-Poly1305
This adds support for XChaCha20-Poly1305, a variant of ChaCha20-Poly1305
with double the nonce size (12->24). Due to the construction of most AEAD
cipher modes which lack a synthetic IV (mostly this and GCM), nonce
reuse becomes problematic. A small default nonce size (96 bits) results
in only 96/2 = 2^48 message encryptions before reuse becomes likely
due to the birthday paradox. As a result, NIST has mandated 2^32 as the
upper limit on AES-GCM encryption operations in FIPS to keep the
threshold sufficiently small. This necessitates key rotation when the
limit is reached.

By using a larger nonce (24-bytes), key rotation can be avoided.

Resolves: https://github.com/openbao/openbao/issues/35

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>