Remove non-OSI database plugins
In alignment with the plugin policy decisions, this removes all non-OSI
database plugins, preferring to make them external plugins.

To ease maintenance burden, only databases licensed under OSI licenses
will be included in the main repository; other plugins will be community
supported and not coupled with the binary or the release process.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>