Trim plugins in openapi.sh
openapi.sh is used to generate the OpenAPI data; to function, it needs
to know a list of all plugins supported by OpenBao in order to
accurately map the dynamically generated APIs.

Note that this does mean that OpenAPI-based data will differ based on
the numbers and types of loaded plugins. Our OpenAPI generated clients
will reflect that supported by the core OpenBao, but if we wish to
generate additional APIs for external-but-supported plugins, we'll have
to re-add them here and modify the script.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>