Move testIdentityStore to use AppRole instead of GitHub
This changes several tests to use the AppRole engine instead of the
GitHub auth plugin. With the removal of non-OSI auth plugins, we needed
a replacement; AppRole is a commonly used plugin (and thus, should be
well tested) to serve as sanity for other parts of OpenBao.

Cherry-pick of f988add19268de52642b1f05bfd222e605d29c64
             + 5c39814b130ac8969525449223d6d2f45e3b3eb9.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>