Remove patch for SHA-1 certificates
Vault < 1.12 supported a patch to allow certificates with SHA-1 hashes
to continue to function after Go removed support in Go 1.18 (or, rather,
gated it behind a feature flag).

As our version number remains unclear, we'll want to remove this patch
in case we choose a lower version number for our releases.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>