Forbid v1 & v2 convergent keys
This prohibits the usage of the nonce field entirely, removing it as an
API option from Transit. Similarly in the keysutil backend, v1 & v2
keys are forbidden.

Because this completely removes nonce from the APIs, this results in
OpenBao issuing a warning about the unknown parameter via the CLI, but
does not fail requests. A nil value for nonce will be passed into the
SDK, which will then generate a fresh value (for encryption) or use the
value inferred from the ciphertext (for decryption).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>