Rename Vault->OpenBao in API responses
This starts to clean up the OpenAPI spec by renaming most instances of
Vault with OpenBao, outside of references to environment variables or
headers.

Note that this will need to be re-done if #64 is accepted and plugins
are imported or forked. When the API has stabilized again, we can
regenerate the OpenAPI spec and check it in.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>