Add ListPageWithContext to API
Unlike most other endpoints, List does not take a generic Data
parameter. For pagination, we wish to standardize the parameters
(generally) and so use the same calling convention (for after and
limit) as used in the storage APIs.

In the future, if more endpoints offer paginated lists and wish to do
so via other filtering mechanisms (or needing multi-part filters for
nested references such as with PKI's revocation queue list endpoints),
a generic ListWithData helper could be added instead.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>