Use ListPage for PKI list endpoints
This uses ListPage for most PKI list endpoints, with the exception of
the complex (two-level: cluster + serial) endpoints for cross-cluster
revocation queues. These hopefully drain rather quickly and it would be
harder to adapt truncation for two parameters (and the CLI would likely
have to allow general K/V parameters to list endpoints which it
currently doesn't as one of the two after values would have to be named
differently).

However, this enables it for leaf and revoked certs (both very large),
roles (which, due to the ACL model, can grow large), and the potentially
less-large but still useful to have paginated on the upper end issuers
and keys listings.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>