Add stub implementation for external plugins
As mentioned in the plugin proposal, external plugins which build
against OpenBao's SDK but run against HashiCorp Vault will need a shim
module to provide compatibility of the ListPage(...) call if it is used.

While the plugin could do this itself (catching the error and then
supplementing its own shim), doing this at the GRPC layer allows plugins
to mostly ignore the differences between the two server versions.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>