Simplify logic in ReadBaoVariable, add LookupBaoVariable
This also enables usage with non-Bao prefixed variables and prevents
weird reads if one is accidentally used. Further, by using
os.LookupEnv(...), we can return an empty BAO_ prefixed variable
value if it is explicitly set, over returning a populated VAULT_
variable instead.

Similarly, we introduce LookupBaoVariable to handle places where
we need to check if the variable was set as well.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>