Switch to running external tools directly
Rather than expecting them to be installed locally (via a previous
`go get` operation outside of the repo), use the `go run ...@latest`
pattern. This will download the package (the first time), but not
necessarily install it for the user on a $PATH-visible location.

This has the added side effect of ensuring everyone is on the same
version.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>