Remove activity log from OpenBao
The activity log subsystem was an upstream specific mechanism to handle
counting of "clients" (for some definition there-of, mostly focused
around combining the concept of an entity with some equivalence relation
for non-entity tokens to reduce the total number), for the express
purpose of billing customers. This code was thus in OSS so that, on
purchasing an Enterprise license, an idea of pricing could be
established.

As we don't have Vault Enterprise, removing this code is likely best as
it allows us to avoid having to maintain it going forward and avoids any
incompatibilities with upstream if they were to change the definitions.

See also: https://developer.hashicorp.com/vault/docs/concepts/client-count/counting
See also: https://developer.hashicorp.com/vault/api-docs/system/internal-counters

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>