Add soft deletion of Transit keys
This adds soft deletion of Transit keys, allowing users to mark keys as
deleted without impacting the ability to backup the key or affecting its
availability to be restored (but preventing all other operations). This
explicitly allows non-exportable keys to be more safely removed, first
via soft deletion (to see if any workloads break and to generally
prevent usage of the key) and then via permanent deletion.

Resolves: #88

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>