Fix handling of JWT audience strings
When aud is a single string (and not a slice of one string) as per
RFC 7519 Section 4.1.3, the JWT auth method ignores this audience,
treating the token as if it had no audience. This is inconsistent
behavior and can result in a auth token being issued when (if encoded
differently) one wouldn't be issued.

This also adds a test validating that multiple audiences are handled
correctly as well.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>