Fix rejection of unauthorized X-Forwarded-For
When x_forwarded_for_reject_not_authorized=false is explicitly set in
the listener configuration, we need to remove the header before
processing the request, otherwise downstream consumers of the request
will not know if the header is authorized or not, and presume that,
because we passed it along, we trusted it.

This is likely not a security issue as plugins do not have access to
X-Forwarded-For by default; only if it were added to the
passthrough_request_headers would it be accessible to plugins.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>