Fix anchor generation in website partials
We use a custom plugin from upstream to include partials in our MDX docs
which can also contain MDX syntax that needs to be evaluated. The
inclusion needs to happen before the default docusaurus remark plugins
run for them to be able to properly evaluate the contained MDX syntax.

One example where this currently does not work is when we overwrite the
generated anchors of a heading.

Signed-off-by: Jan Martens <jan@martens.eu.org>