Correctly seek in raft's ListPage when after=.
When a non-empty path is given that ends in a slash and after=., the
resulting seek prefix from joining the path and the period has the
slash trimmed off. This results in no list results, because this seek
prefix lacks the trailing slash and is thus not prefixed by the expected
prefix. E.g., for foo/ and after=., we end up with seekPrefix=foo, which
doesn't start with our prefix (foo/).

Correctly handle this edge case. This was caught by randomized testing
across multiple storage backends.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>