Set DisableSSCTokens=true as the default config
This changes the default behavior of OpenBao to disable SSCTs, modifying
tests &c to align with this. Internal state required for SSCTs does not
exist in OpenBao (as it is a Vault Enterprise feature) and thus while
existing SSCT wrapped tokens from Vault would work in OpenBao, they are
not populated with consistency information and so no benefit is seen
from keeping them enabled by default.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>