Fix tests for release and remove legacy build process (#362)
* Fix Vault plugin compatibility tests

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Remove upstream HashiCorp Vault build process

This removes unnecessary pipeline executions as we've replaced the
upstream release tooling with goreleaser.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Fix panic in external binary testing

When reverting SSCTs in 15c4855cf6c152dce4828292ec4b4b76c3206f37, this
did not correctly handle making the config DisableSSCTs value a pointer,
resulting in a failure when running with external (Docker) binaries and
potentially full (non-dev-mode) server instances.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Fix ACME EAB test with Caddy

When using `caddy start` with RunCmdWithOutput(...), the container runner
prematurely terminates the backgrounded Caddy server when the latter
function call returns. This prevents the subsequent curl from working as
the server has already been shut down. Switch to running the interactive
variant (`caddy run`) in the background explicitly via
RunCmdInBackground(...).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Fix Vault reference on missing OpenBao UI page

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Fix log statement in goroutine after text exit

Go tests may not log after the end of the test as this will panic the
logger.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

---------

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>