Update Go version, changelog, modules for GA (#410)
* Update Go version, changelog, modules for GA

This introduces a new Makefile target, release-changelog, which requires
to environment variables (LAST_RELEASE, THIS_RELEASE) to generate a
changelog difference between. This can be revised and placed in
CHANGELOG.md and on the website.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

* Fix known vulnerable packages

This updates:

 - github.com/hashicorp/go-retryablehttp
 - github.com/Azure/azure-sdk-for-go/sdk/azidentity
 - github.com/docker/docker@v24.0.9+incompatible

to fix vulnerabilities in these packages.

In updating go-retryablehttp, we pulled in commit
a1a8ab82eb1779b8e09b2d6d2605bbf6fd059a17 which changes the cert
detection behavior. This behavior is not consistent between Go 1.19 and
Go 1.20 (when pulling newer go-retryablehttp versions), so the test is
malformed. Opt to remove it instead.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>

---------

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>