Remove SealUnwrapper
SealUnwrapper exists to err about storage which was at one point in time
seal wrapped. This only exists on Vault Enterprise, and would require a
migration from Vault Enterprise -> OpenBao to trigger. As we likely
won't support this (as we have no way of unsealing the underlying
storage as we don't have the seal wrap subsystem in OSS), and thus
would require it to have been completely unseal wrapped manually
prior to migration, we can safely remove this and assume that other
failures would instead catch this unlikely event.

This is also one less layer to add transactional storage to.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>