Support transactions in AES-GCM Barrier
This extends SecurityBarrier into a TransactionalSecurityBarrier
variant, in which access to the underlying storage (but not the barrier
keyring itself!) can be modified in a transaction-aware way.

Note that modifying the barrier keyring within a transaction makes
little sense: keyring operations are globally locked and we don't want
to rotate to a key used only within a transaction and forked into a
different series later. Thus we only support modifications to the
barrier keyring outside of the scope of a transaction.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>