fix: variable name collision in docker-entrypoint
In the `docker-entrypoint.sh` script there is an issue with collision
between the variable name `BAO_CLUSTER_ADDR` which is used for the
redirect interface and for local JSON configuration. Any attempt to set
JSON configuration will cause `bao` to fail when attempt to parse what
it expects is the cluster interface when it actually JSON configuration.

Therefore, in the lines relating to local JSON configuration the
variable has been renamed to `BAO_LOCAL_CONFIG` allowing for users to
provide custom JSON configuration without also breaking the cluster
address.

Closes #445

Signed-off-by: Jack Hodgkiss <identity@jackhodgkiss.uk>