Add transactions to Raft backend (#436)
This implements transactions in the Raft backend, using an augmented
ApplyLog operation entry. By adding list and get operations (hashing
their output state) to the operation log, we can verify state hasn't
been modified in parallel by another writer, failing the transaction
otherwise.

The current implementation is perhaps inefficient at times: multiple get
or list operations over the same path could potentially be collapsed,
and all get operations moved to the start of transaction (combing the
hashing of multiple entries into a single value, reducing the size of
the log on the wire), but this needs to be done carefully. Thus, this
work is deferred until later when the current implementation is deemed
insufficient.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>