Correctly split TxErrors
This adds a semgrep rule to prevent other instances of the same mistake
in the future. The extent of this issue is that

  errors.Is(err, physical.ErrTransactionCommitFailure) == false

for actual transaction commit failures, though no code currently
conditions around this. The ideal would be that, if an operation
should be retried in the event of a transaction commit failure, it
theoretically could be. But, other errors during the commit process
(not related to consistency) could potentially return a different
error and not trigger an automatic retry.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>