Add in-memory transactional database
This modifies the in-memory physical backend to support transactions in
an inefficient manner: clone the entire tree, build an action log, and
reply the log against the main tree, validating the state of any touched
entries.

A copy-on-write version of go-radix would improve the performance
substantially (letting the old version be maintained in the
transaction), but since this backend isn't persistent across restarts of
OpenBao, it likely is not worth the effort.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>