Prefer openbao_ table prefix for Postgres Storage
This allows an OpenBao and Vault instance to coexist in the same
database at the same time, without trampling each other. In order to do
so, the upsert function (openbao_kv_put) needs to be dynamically
specified, in order to write into the correct table. Using the Vault
version will not work for co-existing in the same database.

This means that any users wishing to in-place migrate from
Vault->OpenBao using the Postgres backend must modify their OpenBao
config to use the old table names if they weren't already specified.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>