Re-add Postgres storage engine (#467)
* Resurrect original Postgres backend

This re-adds the PostgreSQL backend from its state as of the
before-plugin-removal tag (commit
b1f5c1961a6421d50291f16f5cbee6870a5a529c). No changes were made in this
commit from that commit; this can be verified with:

> git diff before-plugin-removal...HEAD -- ./physical/postgresql

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add paginated list to Postgres

This adds support for Paginated lists to Postgres, fixing compatibility
with the OpenBao storage model. This introduces two forms of the
paginated list query, one with a limit and one without, as Postgres's
injection does not allow for the keyword ALL to be used as it is not an
integer (even when passing a string type), nor are negative numbers are
allowed.

We split off SetupDatabaseObjects into a testing.go file, so other test
(which wish to use the Postgres backend) can do so.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Restore PostgreSQL storage documentation

This restores the documentation from
709faf549b73871e99686892a00877310da6f8f7, updating on environment
variable from the VAULT->BAO prefix.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Update VAULT_PG_CONNECTION_URL to use BAO_ prefix

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Re-add PostgreSQL backend to supported server list

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Prefer openbao_ table prefix for Postgres Storage

This allows an OpenBao and Vault instance to coexist in the same
database at the same time, without trampling each other. In order to do
so, the upsert function (openbao_kv_put) needs to be dynamically
specified, in order to write into the correct table. Using the Vault
version will not work for co-existing in the same database.

This means that any users wishing to in-place migrate from
Vault->OpenBao using the Postgres backend must modify their OpenBao
config to use the old table names if they weren't already specified.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Enhancements to PostgreSQL testing

This adds the ability to create the correct ha_table by name and spin up
an on-demand PostgreSQL container for testing with this storage backend.
The latter will be useful for cross-testing contained in the
transactional storage PR.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>