Use transactions for intermediate import
After root CA generation, the next most risky CA operation is
intermediate import. Intermediate signing is usually done on a separate
mount and so doesn't involve as many storage operations as say, root
generation (which also imports the generated root and performs a full
CRL rebuild). However, since this code is shared by the generalized CA
import, and users may be recovering complex chains, it is useful to
protect everything inside a transaction.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>