Update Docker dependency
Docker is not used in the release binary:

    $ go tool nm bao | grep -i '\(docker\|moboy\)'
    $

But it is still worthwhile to update the version pinned in our release
go.mod file anyways.

mholt/archiver and aws/aws-sdk-go are both still vulnerable without
fixes. The former is limited to scenarios when running against a
server (with authentication) to build a debug package (which
occurs locally on the client).

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>