Add transaction support to cache (#442)
Cache is a layer over the top of the physical storage backend (but below
barrier encryption) to reduce strain of repeated read operations on the
backend.

We implement transactions by wrapping the underlying physical
transaction with a new (empty) cache. This means that we may not have
the data locality in oft-read paths, but means that the cache is
internally consistent (and that the transaction's cache doesn't need
to be merged back into the primary cache). In the future, this could be
improved.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>