Add transactions to PKI (#498)
* Use transactions for PKI root generation

This updates PKI root generation to use transactions, giving us an
initial test of the capabilities. PKI root generation involves
modification of several entries, including regenerating all CRLs in
the backend and reading and setting several concurrently-modifiable
config entries, which makes it a good candidate for this.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Use transactions for intermediate import

After root CA generation, the next most risky CA operation is
intermediate import. Intermediate signing is usually done on a separate
mount and so doesn't involve as many storage operations as say, root
generation (which also imports the generated root and performs a full
CRL rebuild). However, since this code is shared by the generalized CA
import, and users may be recovering complex chains, it is useful to
protect everything inside a transaction.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>