Replace more Vault references with OpenBao (#455)
* Replace references of Vault->OpenBao in the CLI

This modifies the CLI to replace Vault with OpenBao in more non-critical
output messages (e.g., help text). While we only guarantee API
compatibility, we also don't want to break scripts unnecessarily.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Finish removing Enterprise-only Transform CLIs

Transform is a separate, internal plugin which isn't built externally
and thus isn't supported on OpenBao. Remove support for it in the CLI.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>