Add transactions to plugin GRPC layer
This adds support for transactions to the plugin GRPC layer,
opportunistically upgrading the client storage implementation to a
transactional aware one if the underlying storage implementation on the
server supports it. This retains bidirectional interoperability with
upstream, allowing plugins compiled with OpenBao's SDK to continue to
work with upstream server implementations and visa-versa.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>