Add transactions to PostgreSQL backend
This implements transaction support in the PostgreSQL backend. Notably,
transactional semantics differ between this and Raft: PostgreSQL will
block on lock acquisition whereas Raft will allow both operations to
continue and fail one of them. This locking property means that the
transactional test suite needed to be reworked first, as it involved
sequential requests to different transactions from the same thread,
causing a stall with PostgreSQL.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>