Replace github.com/mholt/archiver/v3 with stdlib (#611)
* Replace github.com/mholt/archiver/v3 with stdlib

As suggested by Wojciech Kocjan, replace insecure archiver (which lacks
a patched release except via third-party fork) with the standard
library's archive/tar and compress/gzip. This removes it as a direct
dependency.

Notably, our use of archiver was not affected by the vulnerability, as
we only created archives (in `bao debug -compress`) and did not read
them (except for self-created archives as part of the test suite).

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>