Add transactional storage to PostgreSQL (#608)
* Rework transactional testing to be parallelized

The existing testing framework relied on serializable transactions; as
mentioned earlier, this cannot work with PostgreSQL. We leave crosstest
in place as it serves a purpose for comparing various compatible
implementations (like Raft and inmem) plus various higher-level Storage
layers.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add transactions to PostgreSQL backend

This implements transaction support in the PostgreSQL backend. Notably,
transactional semantics differ between this and Raft: PostgreSQL will
block on lock acquisition whereas Raft will allow both operations to
continue and fail one of them. This locking property means that the
transactional test suite needed to be reworked first, as it involved
sequential requests to different transactions from the same thread,
causing a stall with PostgreSQL.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add PostgreSQL-specific clarifications to RFC

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>