Allow PostgreSQL backend to create tables
When running in certain environments such as Kubernetes via Helm chart,
an operator may not necessarily have direct access to the database to
create tables manually. Allow OpenBao to create the tables, which is
enabled by default if it detects it does not have its tables present.

In the future, this sets us up to handle potential schema migrations,
potentially allowing us to scale across multiple tables based on
namespaces or other logical shards.

Resolves: #612

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>