Allow templating policies from JWT claims
This allows users of the JWT auth method to template token policies
based off of the claims in the JWT itself. While not a full replacement
for glob-based matches, this does give the controller some more advanced
control over which policies get added to a template.

By default, we err on missing claims; however, a caller could work
around this by using a complex template such as:

    {{ if ne nil (index . "optional_claim") }}{{ .optional_claim }}{{ end }}

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>