Refactor auto-tidy for correctness
This refactors auto-tidy so that it removes expired, non-revoked
certificates based on safety_buffer but expired, revoked certificates
on revoked_safety_buffer, triggering a CRL build regardless. Revocation
entries are only removed if tidy_revoked_certs=true, but the actual
certificates are removed regardless of tidy_revoked_certs value.

This fixes a bug wherein revoked, expired certificates would be removed
if tidy_cert_store=true and tidy_revoked_certs=true, but safety buffer
was not yet elapsed.

It also unifies CRL rebuild logic, which only occurs when CRLs are not
set to auto-rebuild.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>