Compute raft peer bootstrap challenge via HKDF
Rather than storing a map of serverID->answer (and generating fully
random challenge answers), use HKDF on the serverID and a secret key
to create challenge answers using only a single constant-space key.
HKDF should be at least as fast as the underlying RNG source and this
ensures we do not incur any additional memory per request to this
un-authenticated endpoint.

See also: https://discuss.hashicorp.com/t/hcsec-2024-26-vault-vulnerable-to-denial-of-service-through-memory-exhaustion-when-processing-raft-cluster-join-requests/71047

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>