Clone entity on invalidations
While the invalidation should only be triggered when storage is
writable, if we have an alias invalidation on an entity, we should
only modify the in-memory entity on upsert, not when the alias is
added to the entity. This ensures failures to upsert do not cause a
modified entity to be present in memory.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>