Revert to github.com/hashicorp/raft-boltdb/v2@v2.0.0-20210421194847-a7e34179d62c
This earlier commit lacks metrics (which we do not emit as they were
introduced after us), but also lacks the extra original (archived) bbolt
dependency that causes RISC-V builds to fail. This dependency was
included because subsequent commits added support for migrating from the
archived bbolt to the supported etcd fork, which we do not need as we
never relied on the earlier fork.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>