Add SCAN as logical operation
This introduces SCAN as a new logical operation type, either via a GET
with ?scan=true or via the SCAN HTTP verb. The intent of SCAN is to
allow recursive listing of entries, perhaps using logical.ScanView as a
helper. We require a separate operation because we want to permission
this distinctly from a LIST: operators may wish to deny users a
recursive LIST but grant them a shallow LIST, e.g., for resource usage
or to segment secrets in various other ways.

Notably, with paginated listing, this can still be rather efficient and
need not return all results.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>