Add scan operation for recursive listing (#763)
* Add SCAN as logical operation

This introduces SCAN as a new logical operation type, either via a GET
with ?scan=true or via the SCAN HTTP verb. The intent of SCAN is to
allow recursive listing of entries, perhaps using logical.ScanView as a
helper. We require a separate operation because we want to permission
this distinctly from a LIST: operators may wish to deny users a
recursive LIST but grant them a shallow LIST, e.g., for resource usage
or to segment secrets in various other ways.

Notably, with paginated listing, this can still be rather efficient and
need not return all results.

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add Scan support to KV

Using the new operation, this adds recursive listing (scan) support to
KVv2, allowing a user to see all entries in a single call. With future
support for a paginated form of ScanView, this can be constrained even
in large K/V entries.

Resolves: #549

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add Scan(...) support to the API

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add support for scanning to CLI

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add changelog entry

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add RFC to website

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

* Add scan operation to policy docs

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>

---------

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>