Add pagination_limit to ACL policies
This takes an integer value to limit the maximum value of the page size
(`limit` parameter) in paginated listing requests. This applies to LIST
and SCAN requests. When `limit` is added to `required_parameters` in
conjunction with setting the `pagination_limit` value, the presence of
`limit` is now strictly enforced on requests to this endpoint. However,
`limit` can be silently added if missing from the request (but
`pagination_limit` is present and `required_parameters` does not
include `limit`).

This also adds a helpful value, `max`, which is translated by the ACL
system to the maximum allowed by the operator for this endpoint, or
zero, if no value is defined. This is transparent to the plugin and thus
does not affect typing (it remains an int value).

Resolves: #791

Signed-off-by: Alexander Scheel <ascheel@gitlab.com>